Extract Settings class to separate file from settings window
